requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.1
django-appconf==0.6
django-autofixture==0.10.1
django-autoslug==1.8.0
django-compressor==1.4
django-debug-toolbar==1.2.2
django-extensions==1.5.5
django-jsonfield==0.9.15
#django-toolbelt==0.0.1
django-oauth-toolkit==0.9.0
djangorestframework==3.0.0
drf-nested-routers==0.9.0
ecdsa==0.13
Fabric==1.10.2
gnureadline==6.3.3
gunicorn==19.1.1
httpie==0.9.2
#ipython==3.2.1
#mongoengine==0.10.0
#MySQL-python==1.2.5
#paramiko==1.15.2
Pillow==2.9.0
#psycopg2==2.5.4
pycrypto==2.6.1
Pygments==2.0.2
#pymongo==3.0.3
requests==2.7.0
six==1.8.0
sqlparse==0.1.13
static3==0.5.1
#Werkzeug==0.10.4
wheel==0.24.0
whitenoise==3.2.1